feat: remove errors from infallible sdk functions
Either:

1. These functions statically cannot fail (
https://github.com/filecoin-project/fvm-specs/issues/580 to follow up on
this).
2. These functions can only fail when passed invalid data that we've
ensured is statically impossible with the type system.

fixes #190

+++ fvm/src/syscalls/network.rs
@@ -4,8 +4,8 @@ use fvm_shared::sys;
 use super::Context;
 use crate::kernel::{ClassifyResult, Kernel, Result};
 
-pub fn curr_epoch(context: Context<'_, impl Kernel>) -> Result<u64> {
-    Ok(context.kernel.network_epoch() as u64)
+pub fn curr_epoch(context: Context<'_, impl Kernel>) -> Result<i64> {
+    Ok(context.kernel.network_epoch())
 }
 
 pub fn version(context: Context<'_, impl Kernel>) -> Result<u32> {
